Completely revised data flow between blocks
The `samples` module has been removed and is replaced by a new `signal`
module, which provides the new data type `Signal` that is an enum
containing `Signal::Samples` or a `Signal::Event` as variants.

Event handling has been introduced with the trait `EventHandling` and
the structs `EventHandlers` and `EventHandlerGuard` in the `signal`
module.

The macro `impl_block_trait` simplifies implementation of `Consumer`,
`Producer`, and `EventHandling` traits for signal processing blocks.

The `blocks::io::raw` module has been completely removed in favor of
manually creating `Sender`s and/or `Receiver`s where needed.